requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
bzip2=1.0.8=h5eee18b_5
ca-certificates=2023.12.12=h06a4308_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
ncurses=6.4=h6a678d5_0
openssl=3.0.13=h7f8727e_0
paho-mqtt=2.0.0=pypi_0
pip=23.3.1=py311h06a4308_0
python=3.11.8=h955ad1f_0
readline=8.2=h5eee18b_0
setuptools=68.2.2=py311h06a4308_0
sqlite=3.41.2=h5eee18b_0
tk=8.6.12=h1ccaba5_0
tzdata=2024a=h04d1e81_0
wheel=0.41.2=py311h06a4308_0
xz=5.4.6=h5eee18b_0
zlib=1.2.13=h5eee18b_0
